+++ tasks/send-pan/src/index.ts
@@ -62,6 +62,7 @@ async function sendPAN(event: HandlerEvent): Promise<HandlerOutput> {
       break;
     }
     case 'longPan': {
+      // This assumes one granule per ingest workflow
       if (input.failed.length + input.completed.length <= 1) {
         const disposition = (input.failed.length > 0) ? 'FAILED' : 'SUCCESSFUL';
         pan = pdrHelpers.generateShortPAN(disposition);