+++ CHANGELOG.md
@@ -9,9 +9,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - **CUMULUS-3992**
+  - Update `MoveCollectionsWorkflow` references to `ChangeCollectionsWorkflow`
   - Update `@cumulus/api-client` to add bulkChangeCollection endpoint
-  - Update `@cumulus/api` to add api endpoint to allow trigger of `MoveGranuleCollectionsWorkflow`
-  - Update MoveGranuleCollections integration test to use endpoint to trigger test instead of direct step function invocation
+  - Update `@cumulus/api` to add api endpoint to allow trigger of `ChangeCollectionsWorkflow`
+  - Update ChangeGranuleCollections integration test to use endpoint to trigger test instead of direct step function invocation
 - **CUMULUS-3751**
   - Added `change-granule-collection-s3` to move granules to a different collection.
     - expects a list of granuleIds along with a new (target) collection